+++ src/Controller/Uangkuliah_controller.java
@@ -26,6 +26,7 @@ import javax.swing.table.DefaultTableModel;
  * @author Andi
  */
 public class Uangkuliah_controller {
+
     UangkuliahView view;
     Uangkuliah_model model;
     UangkuliahDao dao;
@@ -40,18 +41,18 @@ public class Uangkuliah_controller {
         dao = new UangkuliahDao();
         mdao = new MahasiswaDao();
         pdao = new ProdiDao();
-         k = new Koneksi();
+        k = new Koneksi();
         try {
             con = k.getConnection();
         } catch (Exception e) {
-            JOptionPane.showMessageDialog(view, "Error: "+e.getMessage());
+            JOptionPane.showMessageDialog(view, "Error: " + e.getMessage());
         }
-        
+
     }
-    
-    String getSelectedButton(){
+
+    String getSelectedButton() {
         String isi = null;
-        for (Enumeration<AbstractButton> buttons = view.getKelompokUkt().getElements();buttons.hasMoreElements();) {
+        for (Enumeration<AbstractButton> buttons = view.getKelompokUkt().getElements(); buttons.hasMoreElements();) {
             AbstractButton button = buttons.nextElement();
             if (button.isSelected()) {
                 isi = button.getText();
@@ -59,26 +60,26 @@ public class Uangkuliah_controller {
             }
         }
         return null;
-    }    
-    
-    public void rbmouseClicked(){
+    }
+
+    public void rbmouseClicked() {
         if (view.getLevel0().isSelected()) {
             getGolongan(view.getLevel0().getText());
         } else if (view.getLevel1().isSelected()) {
             getGolongan(view.getLevel1().getText());
-        }else if (view.getLevel2().isSelected()) {
+        } else if (view.getLevel2().isSelected()) {
             getGolongan(view.getLevel2().getText());
-        }else if (view.getLevel3().isSelected()) {
+        } else if (view.getLevel3().isSelected()) {
             getGolongan(view.getLevel3().getText());
-        }else if (view.getLevel4().isSelected()) {
+        } else if (view.getLevel4().isSelected()) {
             getGolongan(view.getLevel4().getText());
         } else {
             getGolongan(view.getLevel5().getText());
         }
     }
-    
+
     public void getGolongan(String isi) {
-        switch(isi){
+        switch (isi) {
             case "0":
                 view.getLevel0().setSelected(true);
                 view.getSpinJumlah().setValue(0);
@@ -102,11 +103,11 @@ public class Uangkuliah_controller {
             case "5":
                 view.getLevel5().setSelected(true);
                 view.getSpinJumlah().setValue(5000000);
-                break;       
+                break;
         }
-    }    
-    
-    public void clear(){
+    }
+
+    public void clear() {
         view.getTxtNoPembayaran().setText(null);
         view.getTxtNoBp().setText(null);
         view.getTxtNama().setText(null);
@@ -116,8 +117,8 @@ public class Uangkuliah_controller {
         view.getSpinSemester().setValue(1);
         view.getTxtTA().setText(null);
     }
-    
-    public void insert(){
+
+    public void insert() {
         model = new Uangkuliah_model();
         model.setNoPembayaran(view.getTxtNoPembayaran().getText());
         model.setNoBp(view.getTxtNoBp().getText());
@@ -126,18 +127,18 @@ public class Uangkuliah_controller {
         model.setJumlah(view.getSpinJumlah().getValue().toString());
         model.setSemester(view.getSpinSemester().getValue().toString());
         model.setTahunAjaran(view.getTxtTahunAjaran().getText());
-        
+
         try {
             dao.create(model);
             JOptionPane.showMessageDialog(view, "Inserted");
             clear();
         } catch (HeadlessException | SQLException e) {
-              JOptionPane.showMessageDialog(null, 
-                    "ERROR: "+e.getMessage());
+            JOptionPane.showMessageDialog(null,
+                    "ERROR: " + e.getMessage());
         }
     }
-    
-    public void update(){
+
+    public void update() {
         model = new Uangkuliah_model();
         model.setNoPembayaran(view.getTxtNoPembayaran().getText());
         model.setNoBp(view.getTxtNoBp().getText());
@@ -146,28 +147,34 @@ public class Uangkuliah_controller {
         model.setJumlah(view.getSpinJumlah().getValue().toString());
         model.setSemester(view.getSpinSemester().getValue().toString());
         model.setTahunAjaran(view.getTxtTahunAjaran().getText());
-        
+
         try {
-            dao.update(model);
-            JOptionPane.showMessageDialog(view, "Updated");
+            int a = JOptionPane.showConfirmDialog(null, "Yakin mau diubah?", "Konfirmasi", 0, 2);
+            if (a == 0) {
+                dao.update(model);
+                JOptionPane.showMessageDialog(view, "Updated");
+            }
             clear();
         } catch (HeadlessException | SQLException e) {
-              JOptionPane.showMessageDialog(null, 
-                    "ERROR: "+e.getMessage());
+            JOptionPane.showMessageDialog(null,
+                    "ERROR: " + e.getMessage());
         }
     }
-    
-    public void delete(){
+
+    public void delete() {
         String noPembayaran = view.getTxtNoPembayaran().getText();
         try {
-            dao.delete(noPembayaran);
-            JOptionPane.showMessageDialog(null, "DELETED");
+            int a = JOptionPane.showConfirmDialog(null, "Yakin mau dihapus?", "Konfirmasi", 0, JOptionPane.OK_OPTION);
+            if (a == 0) {
+                dao.delete(noPembayaran);
+                JOptionPane.showMessageDialog(null, "DELETED");
+            }
             this.clear();
         } catch (HeadlessException | SQLException e) {
             JOptionPane.showMessageDialog(null, e.getMessage());
         }
     }
-    
+
     public void isiTable() {
         try {
             DefaultTableModel dtm = (DefaultTableModel) view.getTableUkt().getModel();
@@ -184,16 +191,15 @@ public class Uangkuliah_controller {
                     rs.getString(7),
                     rs.getString(8),
                     rs.getString(9),
-                    rs.getString(10),
-                };
+                    rs.getString(10),};
                 dtm.addRow(data);
             }
         } catch (SQLException e) {
-            JOptionPane.showMessageDialog(null, "Error: "+e.getMessage());
+            JOptionPane.showMessageDialog(null, "Error: " + e.getMessage());
         }
     }
-    
-    public void isiTableMahasiswa(){
+
+    public void isiTableMahasiswa() {
         try {
             DefaultTableModel dtm = (DefaultTableModel) view.getTblMhs().getModel();
             dtm.setRowCount(0);
@@ -203,31 +209,32 @@ public class Uangkuliah_controller {
                     rs.getString(1),
                     rs.getString(2),
                     rs.getString(3),
-                    rs.getString(4),
-                };
+                    rs.getString(4),};
                 dtm.addRow(data);
             }
         } catch (SQLException e) {
-            JOptionPane.showMessageDialog(null, "Error: "+e.getMessage());
+            JOptionPane.showMessageDialog(null, "Error: " + e.getMessage());
         }
     }
-    
-    public void tblMhsMouseClicked(){
+
+    public void tblMhsMouseClicked() {
         mdao = new MahasiswaDao();
         String noBp = view.getTblMhs().getValueAt(
-            view.getTblMhs().getSelectedRow(),0).toString();
+                view.getTblMhs().getSelectedRow(), 0).toString();
         try {
             mmodel = mdao.getTbMahasiswa(noBp);
             view.getTxtNoBp().setText(mmodel.getNo_bp());
             view.getTxtNama().setText(mmodel.getNama());
-        } catch (SQLException e) {JOptionPane.showMessageDialog(view, "Error : "+e);}
+        } catch (SQLException e) {
+            JOptionPane.showMessageDialog(view, "Error : " + e);
+        }
     }
-    
-    public void tableUktMouseClicked(){
+
+    public void tableUktMouseClicked() {
         dao = new UangkuliahDao();
         String nopembayaran = view.getTableUkt().getValueAt(
-                view.getTableUkt().getSelectedRow(),0).toString();
-        
+                view.getTableUkt().getSelectedRow(), 0).toString();
+
         try {
             model = dao.getTbUangkuliah(nopembayaran);
             mmodel = mdao.getTbMahasiswa(model.getNoBp());
@@ -239,7 +246,9 @@ public class Uangkuliah_controller {
             getGolongan(model.getGolongan());
             view.getSpinSemester().setValue(Integer.parseInt(model.getSemester()));
             view.getTxtTA().setText(model.getTahunAjaran());
-            
-        } catch (SQLException e) {JOptionPane.showMessageDialog(view, "Error : "+e);}
+
+        } catch (SQLException e) {
+            JOptionPane.showMessageDialog(view, "Error : " + e);
+        }
     }
 }
